refactor cmake, add boost::random for portable randomness
- add boost::random for portable randomness generation. std::uniform_int_distribution
is implementation dependent (e.g., gnu-gcc 9 and clang-12 have different implementations)
We thus had issues with generating reproducible streams on different platforms.
We use Boost::random as a portable distribution sampler, this could be changed later
as interface is now eacirc::uniform_int_distribution.

- it is possible to disable portable randomness with cmake parameter PORTABLE_RANDOMNESS=OFF,
then Boost is not needed. However, this is not recommended.

- fixing git cmake issues: generate version tag even there is no tag in the git, use git
repo of the eacirc-core if used as submodule (not the parent repo)